requirements.txt
azure-ai-ml>=1.17.0
azure-identity>=1.15.0
mlflow>=2.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
